+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 92:
 * Fix typo in test/CMakeLists.txt
 * basic_fields::value_type is not copyable
 * Update repository links in source comments
+* Ignore Content-Length in some cases
 
 --------------------------------------------------------------------------------